Update BugSigDB_paper_to_curate.yml
Adds two unique identifiers:
1. PMID, optional.
2. DOI, required.

The idea is to ensure that a computable unique identifier is attached to studies. While PMID may not be available, I cannot think of any source of literature that would enter the curation process that will not have a DOI. 

The DOI allows for automated metadata population, provides clear provenance for the curation (ie., the curation is based on information from a specific DOI), and is guaranteed to be available.